packages/requirements.txt
# Core NLP - Flair based
flair>=0.12.2
torch>=2.0.0

# Original dependencies
presidio-analyzer>=2.2.33
presidio-anonymizer>=2.2.33
faker>=19.0.0
pydantic>=2.0.0
pytest>=7.0.0

# Additional utilities
regex>=2023.0.0
pyyaml>=6.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
ruff>=0.1.0

# Document processing (Phase 2)
# PDF support
pdfplumber>=0.10.0
PyPDF2>=3.0.0
reportlab>=4.0.0

# OCR support
pdf2image>=1.16.0
pytesseract>=0.3.10
easyocr>=1.7.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Office documents
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21

# Email processing
extract-msg>=0.45.0